+++ src/Modules/Pricing/Hotelio.Modules.Pricing.Application/Command/AddRoomTariff.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace Hotelio.Modules.Pricing.Application.Command;
 
-internal record AddRoomTariff(Guid HotelTariffId, string RoomTypeId, double BasePriceNetAmount) : IRequest;
+internal record AddRoomTariff(Guid HotelTariffId, string RoomTypeId, decimal BasePriceNetAmount) : IRequest;
