+++ src/Modules/Pricing/Hotelio.Modules.Pricing.Application/Query/Calculate.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace Hotelio.Modules.Pricing.Application.Query;
 
-internal record Calculate(string HotelId, string RoomType, List<string> Amenities, DateTime StartDate, DateTime EndDate) : IQuery<double>;
+internal record Calculate(string HotelId, string RoomType, List<string> Amenities, DateTime StartDate, DateTime EndDate) : IQuery<decimal>;
