+++ src/Shared/Hotelio.Shared/Domain/Price.cs
@@ -4,18 +4,18 @@ namespace Hotelio.Shared.Domain;
 
 public record Price
 {
-    public double NetAmount { get; init; }
+    public decimal NetAmount { get; init; }
     public Currency Currency { get; init; }
     public int TaxRate { get; init; }
 
-    private Price(double netAmount, Currency currency, int taxRate)
+    private Price(decimal netAmount, Currency currency, int taxRate)
     {
         NetAmount = netAmount;
         Currency = currency;
         TaxRate = taxRate;
     }
 
-    public static Price Create(double netAmount, Currency currency, int taxRate)
+    public static Price Create(decimal netAmount, Currency currency, int taxRate)
     {
         if (taxRate > 100)
         {
@@ -25,13 +25,13 @@ public record Price
         return new Price(netAmount, currency, taxRate);
     }
     
-    public static Price CreateDefault(double netAmount)
+    public static Price CreateDefault(decimal netAmount)
     {
         return new Price(netAmount, Currency.PLN, 23);
     }
 
-    public double GetGrossAmount()
+    public decimal GetGrossAmount()
     {
-        return NetAmount * (1 + TaxRate / 100.0);
+        return NetAmount * (1 + TaxRate / 100m);
     }
 }
